Add consumer batch
